fix(webchat): address review feedback
- Extract formatDuration to shared util, deduplicate cron pages
- Add debug log for LatestTurnNum DB error in bridge_forward
- Swap findChecksum check order for fast path optimization
