+++ webchat/app/admin/cron/detail/page.tsx
@@ -5,6 +5,7 @@ import { useSearchParams, useRouter } from 'next/navigation';
 import Link from 'next/link';
 import { listCronJobs, updateCronJob, deleteCronJob, triggerCronJob } from '@/lib/api/admin-cron';
 import { useAdminUI } from '@/context/admin-ui-context';
+import { formatDuration } from '@/lib/utils/format-duration';
 import type { CronJob } from '@/lib/types/admin';
 
 // ---------------------------------------------------------------------------
@@ -19,17 +20,6 @@ function formatScheduleStr(s?: CronJob['schedule']): string {
     case 'at': return `at:${s.at ?? ''}`;
     default: return '';
   }
-}
-
-function formatDuration(ms: number): string {
-  const sec = Math.floor(ms / 1000);
-  if (sec < 60) return `${sec}s`;
-  const min = Math.floor(sec / 60);
-  if (min < 60) return sec % 60 ? `${min}m${sec % 60}s` : `${min}m`;
-  const hr = Math.floor(min / 60);
-  if (hr < 24) return min % 60 ? `${hr}h${min % 60}m` : `${hr}h`;
-  const d = Math.floor(hr / 24);
-  return hr % 24 ? `${d}d${hr % 24}h` : `${d}d`;
 }
 
 function InfoRow({ label, value, mono }: { label: string; value: string; mono?: boolean }) {
