+++ webchat/app/admin/cron/page.tsx
@@ -4,6 +4,7 @@ import { useCallback, useEffect, useState } from 'react';
 import Link from 'next/link';
 import { listCronJobs, updateCronJob, deleteCronJob, triggerCronJob } from '@/lib/api/admin-cron';
 import { useAdminUI } from '@/context/admin-ui-context';
+import { formatDuration } from '@/lib/utils/format-duration';
 import type { CronJob } from '@/lib/types/admin';
 import type { CronSchedule } from '@/lib/types/admin';
 
@@ -21,17 +22,6 @@ function formatSchedule(s: CronJob['schedule']): string {
     case 'at': return s.at ?? '—';
     default: return s.kind ?? '—';
   }
-}
-
-function formatDuration(ms: number): string {
-  const sec = Math.floor(ms / 1000);
-  if (sec < 60) return `${sec}s`;
-  const min = Math.floor(sec / 60);
-  if (min < 60) return sec % 60 ? `${min}m${sec % 60}s` : `${min}m`;
-  const hr = Math.floor(min / 60);
-  if (hr < 24) return min % 60 ? `${hr}h${min % 60}m` : `${hr}h`;
-  const d = Math.floor(hr / 24);
-  return hr % 24 ? `${d}d${hr % 24}h` : `${d}d`;
 }
 
 function formatTime(ms?: number): string {
